requirements.txt
streamlit
google-genai
python-dotenv
PyPDF2
python-docx
pyperclip 